fix: move tests to dgram_tests
